move audio deletion functionality to dedicated tab
